+++ staking/package.json
@@ -35,8 +35,7 @@
     "test": "npm run build_wasm && anchor build -- --features mock-clock && npm run dump_governance && ts-mocha --parallel -p ./tsconfig.json -t 1000000 tests/*.ts",
     "build": "npm run build_wasm && tsc -p tsconfig.api.json",
     "build_wasm": "./scripts/build_wasm.sh",
-    "localnet" : "anchor build && ts-node ./app/scripts/localnet.ts",
-    "setup": "anchor build && npm run dump_governance && ts-node ./app/scripts/setup_local.ts",
+    "localnet" : "anchor build && npm run dump_governance && ts-node ./app/scripts/localnet.ts",
     "dump_governance": "./clone_programs.sh",
     "prepare": "cd .. && husky install staking/.husky"
